# packages/service/openapi.json
{
  "components": {
    "schemas": {
...
      "CharsetMeta": {
        "properties": {
          "drawn": {
            "title": "Drawn",
            "type": "integer"
          },
          "encoded": {
            "title": "Encoded",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "encoded",
          "drawn",
          "pages"
        ],
        "title": "CharsetMeta",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "GlyphList": {
        "properties": {
          "glyphs": {
            "items": {
              "$ref": "#/components/schemas/GlyphResponse"
            },
            "title": "Glyphs",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "glyphs"
        ],
        "title": "GlyphList",
        "type": "object"
      },
      "GlyphResponse": {
        "properties": {
          "advance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advance"
          },
          "char": {
            "title": "Char",
            "type": "string"
          },
          "codepoint": {
            "title": "Codepoint",
            "type": "string"
          },
          "status": {
            "enum": [
              "missing",
              "auto",
              "accepted",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "svg_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svg Url"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "enum": [
                "LOW_INK",
                "TOUCHES_BORDER",
                "TINY_CONTOURS_REMOVED",
                "LARGE_INK_BLOB",
                "OFF_GUIDE"
              ],
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "codepoint",
          "char",
          "status",
          "warnings",
          "updated_at"
        ],
        "title": "GlyphResponse",
        "type": "object"
      },
...
      "MetaResponse": {
        "properties": {
          "charsets": {
            "items": {
              "$ref": "#/components/schemas/CharsetMeta"
            },
            "title": "Charsets",
            "type": "array"
          },
          "retention_days": {
            "title": "Retention Days",
            "type": "number"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "retention_days",
          "charsets"
        ],
        "title": "MetaResponse",
        "type": "object"
      },
...
      "ReviewError": {
        "properties": {
          "codepoint": {
            "title": "Codepoint",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "codepoint",
          "reason"
        ],
        "title": "ReviewError",
        "type": "object"
      },
      "ReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "accept": {
            "items": {
              "type": "string"
            },
            "maxItems": 400,
            "title": "Accept",
            "type": "array"
          },
          "reject": {
            "items": {
              "type": "string"
            },
            "maxItems": 400,
            "title": "Reject",
            "type": "array"
          }
        },
        "title": "ReviewRequest",
        "type": "object"
      },
      "ReviewResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ReviewError"
            },
            "title": "Errors",
            "type": "array"
          },
          "unchanged": {
            "title": "Unchanged",
            "type": "integer"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          }
        },
        "required": [
          "updated",
          "unchanged",
          "errors"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
      "UploadResponse": {
        "properties": {
          "deduplicated": {
            "title": "Deduplicated",
            "type": "boolean"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "upload_id": {
            "format": "uuid",
            "title": "Upload Id",
            "type": "string"
          }
        },
        "required": [
          "upload_id",
          "job_id",
          "deduplicated"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "glyphlab API",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/meta": {
      "get": {
        "operationId": "get_meta",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "411": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Length Required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Meta"
      }
    },
...
    "/api/projects/{project_id}/glyphs": {
      "get": {
        "operationId": "list_glyphs",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "missing",
                      "auto",
                      "accepted",
                      "rejected"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "warning",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "LOW_INK",
                    "TOUCHES_BORDER",
                    "TINY_CONTOURS_REMOVED",
                    "LARGE_INK_BLOB",
                    "OFF_GUIDE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warning"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": -1,
              "maximum": 1114111,
              "minimum": -1,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 300,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlyphList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Glyphs"
      }
    },
...
    "/api/projects/{project_id}/glyphs:review": {
      "post": {
        "operationId": "review_glyphs",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Review Glyphs"
      }
    },
...
    "/api/projects/{project_id}/uploads": {
      "post": {
        "operationId": "create_upload",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "OK"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Insufficient Storage"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Upload"
      }
    },
...
  }
}